pyproject.toml
[tool.poetry]
name = "notification-system "
version = "0.1.0"
description = ""
authors = ["baktiyar <baktiarzaksylykov22@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
pydantic = {extras = ["email"], version = "^2.11.7"}
arq = "^0.26.3"
httpx = "^0.28.1"
aiosmtplib = "^4.0.1"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
